requirements.txt
dvc == 1.11.4
matplotlib == 3.3.2
numpy == 1.18.5
plotly == 4.14.1
pyyaml == 5.3.1
scikit_image == 0.17.2
scikit_learn == 0.23.2
tensorflow == 2.3.1
